refactor(shared): extract fetchRootlistPlaylists into @shared/api
