+++ packages/shared/src/api/index.ts
@@ -1,3 +1,4 @@
 export { cosmos } from './cosmos';
 export { resolveUriMetadata, type UriMeta } from './uri-metadata';
+export { fetchRootlistPlaylists, type PlaylistRef } from './rootlist';
 export { paginate, batchedWrite, PAGE_SIZE, BATCH_DELAY_MS, PLAYLIST_BATCH_SIZE } from './batch';
